requirements.txt
bert_score==0.3.13
clip==1.0
cpm_kernels==1.0.11
decord==0.6.0
deepspeed==0.18.2
easydict==1.13
einops==0.8.1
flash_attn==2.8.3
gradio==6.0.1
moverscore==1.0.3
numpy==2.3.5
openai_clip==1.0.1
peft==0.18.0
Pillow==12.0.0
pycocoevalcap==1.2
Requests==2.32.5
sentencepiece==0.2.1
torch==2.4.0
torchvision==0.19.0
tqdm==4.67.1
tqdm==4.66.4
transformers==4.37.2
